refactor(db): extract recommendation types to separate module
Moved recommendation-related interfaces and types from recommendation-engine.ts to a dedicated recommendation-types.ts file to improve code organization and reduce circular dependencies. This change centralizes all recommendation type definitions in one location, making the codebase more maintainable and easier to understand.

The refactoring affects multiple files including the recommendation engine, similar statistics modules, and API routes, which now import types from the new centralized location.
