+++ apps/nextjs-app/app/(app)/servers/[id]/(auth)/dashboard/SimilarSeriesStatistics.tsx
@@ -1,11 +1,13 @@
 "use client";
 
 import { Monitor } from "lucide-react";
+import type {
+  RecommendationSource,
+  SeriesRecommendationItem,
+} from "@/lib/db/recommendation-types";
 import {
   getSimilarSeries,
   hideSeriesRecommendation,
-  type RecommendationSource,
-  type SeriesRecommendationItem,
 } from "@/lib/db/similar-series-statistics";
 import type { ServerPublic } from "@/lib/types";
 import { RecommendationsSection } from "./RecommendationsSection";
