+++ apps/nextjs-app/app/(app)/servers/[id]/(auth)/dashboard/SimilarStatistics.tsx
@@ -1,11 +1,13 @@
 "use client";
 
 import { Film } from "lucide-react";
+import type {
+  RecommendationItem,
+  RecommendationSource,
+} from "@/lib/db/recommendation-types";
 import {
   getSimilarStatistics,
   hideRecommendation,
-  type RecommendationItem,
-  type RecommendationSource,
 } from "@/lib/db/similar-statistics";
 import type { ServerPublic } from "@/lib/types";
 import { RecommendationsSection } from "./RecommendationsSection";
