+++ apps/nextjs-app/app/api/recommendations/route.ts
@@ -4,21 +4,19 @@ import {
   authenticateMediaBrowser,
   validateJellyfinToken,
 } from "@/lib/api-auth";
+import type {
+  RecommendationItem,
+  RecommendationSource,
+  SeriesRecommendationItem,
+} from "@/lib/db/recommendation-types";
 import {
   hasServerIdentifier,
   parseServerIdentifier,
   resolveServer,
   type ServerIdentifier,
 } from "@/lib/db/server-resolver";
-import {
-  getSimilarSeries,
-  type SeriesRecommendationItem,
-} from "@/lib/db/similar-series-statistics";
-import {
-  getSimilarStatistics,
-  type RecommendationItem,
-  type RecommendationSource,
-} from "@/lib/db/similar-statistics";
+import { getSimilarSeries } from "@/lib/db/similar-series-statistics";
+import { getSimilarStatistics } from "@/lib/db/similar-statistics";
 import { authenticateByName } from "@/lib/jellyfin-auth";
 
 type RecommendationType = "Movie" | "Series" | "all";
