+++ apps/nextjs-app/lib/db/similar-series-statistics.ts
@@ -18,18 +18,16 @@ import {
   sql,
 } from "drizzle-orm";
 import { revalidateTag } from "next/cache";
-import {
-  getProfileRecommendations,
-  type ProfileRecommendationsResponse,
-  type RecommendationCardItem,
-  type RecommendationResult,
-  type RecommendationSource,
-} from "./recommendation-engine";
+import { getProfileRecommendations } from "./recommendation-engine";
+import type {
+  ProfileRecommendationsResponse,
+  RecommendationCardItem,
+  RecommendationResult,
+} from "./recommendation-types";
 import { getMe } from "./users";
 
 export type SeriesRecommendationItem = RecommendationResult;
 export type SeriesRecommendationCardItem = RecommendationCardItem;
-export type { ProfileRecommendationsResponse, RecommendationSource };
 
 type RecommendationCardItemWithEmbedding = RecommendationCardItem & {
   embedding: Item["embedding"];
